discogs_credits: extract constants + dom event helpers
Issue #32 — next slice of the multi-file source split.

  src/constants.js  REL_TEMPLATE, SELECTORS, DISCOGS_LOGO_URL  (~55 lines)
  src/util.js       doNext, setNativeValue, selectValue,
                    makeKeyDownEvent, makeClickEvent           (~90 lines)

Entry file now imports these. ~100 lines moved out. No behavior change;
full 12-fixture suite still green.
